Security

+++ src/main/java/by/academy/fitness/service/DishService.java
@@ -6,6 +6,7 @@ import java.util.List;
 import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -14,30 +15,42 @@ import by.academy.fitness.dao.Filtering;
 import by.academy.fitness.dao.Sorting;
 import by.academy.fitness.domain.dto.DishDTO;
 import by.academy.fitness.domain.dto.IngredientDTO;
+import by.academy.fitness.domain.entity.Audit;
 import by.academy.fitness.domain.entity.Dish;
 import by.academy.fitness.domain.entity.Ingredient;
 import by.academy.fitness.domain.entity.Page;
+import by.academy.fitness.domain.entity.User;
+import by.academy.fitness.domain.entity.Audit.ESSENCETYPE;
 import by.academy.fitness.domain.validators.DishValidator;
 import by.academy.fitness.service.interf.IDishService;
 
 @Service
 public class DishService implements IDishService {
+	private final static String CREATED = "New dish created";
+	private final static String UPDATED = "Dish updated";
+	private final static String DELETED = "Dish deleted";
 
 	private final DishDao dishDao;
 	private final ProductService productService;
 	private final DishValidator validator;
+	private final UserService userService;
+	private final AuditService auditService;
 
 	@Autowired
-	public DishService(DishDao dishDao, ProductService productService, DishValidator validator) {
+	public DishService(DishDao dishDao, ProductService productService, DishValidator validator, UserService userService,
+			AuditService auditService) {
 		super();
 		this.dishDao = dishDao;
 		this.productService = productService;
-		this.validator=validator;
+		this.validator = validator;
+		this.userService = userService;
+		this.auditService = auditService;
 	}
 
 	@Transactional
 	@Override
 	public Dish create(DishDTO dto) {
+		User user = userService.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName());
 		Dish dish = new Dish();
 		validator.validate(dto);
 		dish.setUuid(UUID.randomUUID());
@@ -54,6 +67,8 @@ public class DishService implements IDishService {
 			ingr.add(x);
 		}
 		dish.setIngredients(ingr);
+		auditService.create(new Audit(CREATED, ESSENCETYPE.DISH, dish.getName() + " " + dish.getUuid().toString()),
+				user);
 
 		return dishDao.create(dish);
 	}
@@ -66,12 +81,15 @@ public class DishService implements IDishService {
 	@Transactional
 	@Override
 	public Dish update(UUID uuid, LocalDateTime dtUpdate, DishDTO dto) {
+		User user = userService.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName());
 		Dish readed = dishDao.findByUuid(uuid);
 
 		if (readed == null) {
 			throw new IllegalArgumentException("Позиция не найдена");
 		}
-
+		if (!readed.getUser().equals(user)) {
+			throw new IllegalArgumentException("You can only update the product you created");
+		}
 //		if (!readed.getDtUpdate().isEqual(dtUpdate)) {
 //			throw new IllegalArgumentException("К сожалению позиция уже была отредактирована кем-то другим");
 //		}
@@ -87,6 +105,8 @@ public class DishService implements IDishService {
 			ingr.add(x);
 		}
 		readed.setIngredients(ingr);
+		auditService.create(new Audit(UPDATED, ESSENCETYPE.DISH, readed.getName() + " " + readed.getUuid().toString()),
+				user);
 
 		return dishDao.create(readed);
 	}
@@ -94,6 +114,11 @@ public class DishService implements IDishService {
 	@Transactional
 	@Override
 	public void delete(UUID uuid, LocalDateTime dtUpdate) {
+		User user = userService.findByEmail(SecurityContextHolder.getContext().getAuthentication().getName());
+		if (!dishDao.findByUuid(uuid).getUser().equals(user)) {
+			throw new IllegalArgumentException("You can only delete the product you created");
+		}
+		auditService.create(new Audit(DELETED, ESSENCETYPE.DISH, uuid.toString()), user);
 		dishDao.delete(uuid, dtUpdate);
 
 	}
@@ -101,26 +126,27 @@ public class DishService implements IDishService {
 	@Transactional
 	@Override
 	public Page<Dish> get(Integer amount, Integer skip, List<Sorting> sortings, List<Filtering> filters) {
-			Page<Dish> page = new Page<>();
-			page.setContent(dishDao.findAll(amount, skip, sortings, filters));
-			page.setPageSize(amount);
-			int count = dishDao.count(filters);
-			page.setTotalElements(count);
-			int pageSize = count / (amount.intValue() == 0 ? 1 : amount.intValue());
-			if (count % (amount.intValue() == 0 ? 1 : amount.intValue()) > 0) {
-				pageSize++;
-			}
-			int currentPage = skip.intValue()/amount.intValue();
-			page.setPageNumber(currentPage);
-			page.setTotalPages(pageSize);
-			page.setFirst(skip == 0);
-			page.setLast((count-skip)<=amount);;
-			int numberOfEl = amount;
-			if (amount>(count-skip)) {
-				numberOfEl=count-skip;
-			}
-			page.setNumberOfElements(numberOfEl);
-			return page;
+		Page<Dish> page = new Page<>();
+		page.setContent(dishDao.findAll(amount, skip, sortings, filters));
+		page.setPageSize(amount);
+		int count = dishDao.count(filters);
+		page.setTotalElements(count);
+		int pageSize = count / (amount.intValue() == 0 ? 1 : amount.intValue());
+		if (count % (amount.intValue() == 0 ? 1 : amount.intValue()) > 0) {
+			pageSize++;
+		}
+		int currentPage = skip.intValue() / amount.intValue();
+		page.setPageNumber(currentPage);
+		page.setTotalPages(pageSize);
+		page.setFirst(skip == 0);
+		page.setLast((count - skip) <= amount);
+		;
+		int numberOfEl = amount;
+		if (amount > (count - skip)) {
+			numberOfEl = count - skip;
+		}
+		page.setNumberOfElements(numberOfEl);
+		return page;
 	}
 
 }
